+++ RELEASES.md
@@ -1,5 +1,8 @@
 # Astra Monitor 14 - TO BE RELEASED
 
+### New features
+- Added a shortcut to set all main and secondary colors all at once [[#74](https://github.com/AstraExt/astra-monitor/issues/74)]
+
 ### Bug fixes
 - Fixed a bug where a task could stop updating when a cancel was requested